requirements.txt
psycopg2-binary
simplejson
pyotp
qrcode
requests
gunicorn
libmc
libgravatar
Flask
SQLAlchemy
Flask-SQLAlchemy
Flask-Migrate
Flask-Script
Flask-Admin
Flask-Cors
Pillow
APScheduler
python-dotenv

flake8
yapf
